added add multiparte and remove multipare tool

+++ vue/components/toolsoftool/snap.vue
@@ -1,7 +1,9 @@
 <template>
-  <div>
+  <div style="display: flex; justify-content: space-between; width: 100%">
     <input type="checkbox" class="magic-checkbox snap_tools_of_tools" :id="id" v-model="checked">
     <label :for="id">Snap</label>
+    <input type="checkbox" class="magic-checkbox snap_tools_of_tools"  :id="idAll" v-model="checkedAll">
+    <label :for="idAll">Snap All</label>
   </div>
 </template>
 
@@ -9,38 +11,102 @@
   const GUI = g3wsdk.gui.GUI;
   let snapInteraction;
   const mapService = GUI.getComponent('map').getService();
+  const editingService = require('../../../services/editingservice');
   export default {
     name: "snap",
     props: ['options'],
     data() {
       return {
         id: `snap_${Date.now()}`,
-        checked: false
+        idAll: `snap_${Date.now()}_all`,
+        checked: false,
+        checkedAll: false
       }
     },
     computed: {
       add() {
-        return this.checked && this.active;
+        return (this.checked || this.checkedAll) && this.active;
       },
       active() {
         return this.options.active;
       }
     },
+    methods: {
+      addFeatures(features){
+        this.features.extend(features)
+      },
+      addFeature(feature){
+        this.addFeatures([feature]);
+      },
+      removeFeatures(features){
+        features.forEach(feature => this.features.remove(feature));
+      }
+    },
     watch: {
-      'add': function(add) {
-        if (add) {
-          snapInteraction = new ol.interaction.Snap({
+      add(add) {
+        if (!add) {
+          mapService.removeInteraction(snapInteraction);
+        }
+      },
+      checked(checked){
+        if (checked) {
+          this.checkedAll &&  mapService.removeInteraction(snapInteraction);
+          this.checkedAll = false;
+          snapInteraction =   new ol.interaction.Snap({
             source: this.options.source
           });
           mapService.addInteraction(snapInteraction);
-        } else {
-          mapService.removeInteraction(snapInteraction);
-          snapInteraction = null;
+        }
+      },
+      checkedAll(checked){
+        if (checked) {
+          this.checked &&  mapService.removeInteraction(snapInteraction);
+          this.checked = false;
+          snapInteraction = new ol.interaction.Snap({
+            features: this.features
+          });
+          mapService.addInteraction(snapInteraction);
         }
       }
     },
+    created() {
+      this.features = new ol.Collection();
+      this.sourcesAndEventsKeys = [];
+      editingService.getToolBoxes().filter(toolbox => {
+        if (toolbox.getLayer().type === 'vector') {
+          const source = toolbox.getLayer().getEditingSource();
+          this.features.extend(source.readFeatures());
+          const addFeaturesKey = source.onbefore('addFeatures', this.addFeatures);
+          const addFeatureKey = source.onbefore('addFeature', this.addFeatures);
+          const clearKey = source.onbefore('clear', () =>{
+            const features = source.readFeatures();
+            this.removeFeatures(features);
+          });
+          const olKey = source.getFeaturesCollection().on('add', (evt)=>{
+            this.addFeature(evt.element)
+          });
+          this.sourcesAndEventsKeys.push({
+            source,
+            settersAndKeys: {
+              'addFeatures': addFeaturesKey,
+              'addFeature': addFeatureKey,
+              'clear': clearKey
+            },
+            olKey
+          })
+        }
+      })
+    },
     destroyed() {
       snapInteraction && mapService.removeInteraction(snapInteraction);
+      this.sourcesAndEventsKeys.forEach(sourceAndKey =>{
+        const {source, settersAndKeys, olKey} = sourceAndKey;
+        Object.keys(settersAndKeys).forEach(eventName =>{
+          const key = settersAndKeys[eventName];
+          source.un(eventName, key)
+        });
+        ol.Observable.unByKey(olKey)
+      })
     }
   }
 </script>
